+++ core/src/main/scala/org/easysql/database/Page.scala
@@ -1,3 +1,3 @@
 package org.easysql.database
 
-case class Page[T](totalPage: Int = 0, totalCount: Int = 0, data: List[T] = List())
+case class Page[T](totalPage: Long = 0l, totalCount: Long = 0l, data: List[T] = List())
